+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - receiving payments in crypto for subscription
 - validating payments amount in webhook
+- monitoring system with notifications to Bugsnag
 
 ## [1.6.8] - 2024-01-09
 ### Added